Fixing previous functionality
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/branches/dev/Modules/Spatial/APP@3366 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ rest/src/main/java/eu/europa/ec/fisheries/uvms/spatial/rest/resources/EezResource.java
@@ -28,7 +28,7 @@ public class EezResource {
     @Consumes(value = {MediaType.APPLICATION_JSON})
     @Produces(value = {MediaType.APPLICATION_JSON})
     @Path("/{id}")
-    public ResponseDto getExclusiveEconomicZoneById(@PathParam("id") long eezId) {
+    public ResponseDto getExclusiveEconomicZoneById(@PathParam("id") int eezId) {
         try {
             LOG.info("Getting eez with {}", eezId);
             GetEezSpatialRS eez = exclusiveEconomicZoneService.getExclusiveEconomicZoneById(eezId);
