+++ service/src/main/java/eu/europa/ec/fisheries/uvms/spatial/service/bean/ExclusiveEconomicZoneService.java
@@ -7,5 +7,5 @@ import eu.europa.ec.fisheries.schema.spatial.source.GetEezSpatialRS;
  */
 public interface ExclusiveEconomicZoneService {
 
-    GetEezSpatialRS getExclusiveEconomicZoneById(long id);
+    GetEezSpatialRS getExclusiveEconomicZoneById(int id);
 }
